Update the SCH codebase to the more recent cFS revision (March 2023)
This changeset contains 98%+ fixes that are needed to have this SCH
working against the more recent cFS setup as of March 2023.

The cFS revision that this is compiled against: 192f2ea3061bfe722b8ed1d72b7adecbf36d4f6e

https://github.com/nasa/cFS/commit/192f2ea3061bfe722b8ed1d72b7adecbf36d4f6e

Known issues:

- The changeset contains minor formatting issues but I am deciding to
ignore them because the SCH itself needs to be Clang Formatted and I
didn't find an existing .clang-format in the repository.
- This changeset does not touch anything related to setting up the table
files in the CMake build system. There are other patches that are
attempting to fix those issues: https://github.com/nasa/SCH/pull/6/files
- We are still reworking the SCH codebase to fit our needs, so it may be
that a few lines were missed and they will not compile.